cmd/vet: check for programming errors
Command vet is a fork of the go tool 'vet', similar in
spirit but looking for errors that are specific to this
code base.

Initially, there is only one check: vet reports calls to
log.Write with an odd number of key-value arguments.
This command doesn't duplicate the 'go vet' checks. Both
tools must be run to check for all errors.

This command, unlike the go tool 'vet', skips
directories that start with '.' or '_', as well as those
named 'testdata' or 'vendor'.

Closes chain/chainprv#17.
Reviewers: @erykwalder

Closes chain/chainprv#17.
Reviewers: @erykwalder