errors: add detail field
Callers can use WithDetail to carry a specific message
along with the error value, distinct from the underlying
error value and context. This detail message won't
contain the same text returned when the error formats
itself for printing.

This structure can be used to provide more information
to the user when an error occurs. For example, the root
error message might be "object not found" and the detail
might say "loading user u123". Or the message might say
"Redeem script is not a valid multisig script" and the
detail might say "on input 2".

By contrast, the error might contain a great deal more
context that shouldn't be exposed to the end user,
making the formatted error message unsuitable to expose.

Closes chain/chainprv#31.
Reviewers: @jeffomatic