api/appdb: Store keys as xpubs, remove keys table
The notion of an xpub hash was created to keep Chain's own xpubs
private, since they were shared between customers and could be used to
discover information about others' accounts.

Now that Chain no longer acts as a signer, it is much more direct to
refer to keys by their xpub, rather than by their xpub hash. This
eliminates the need for the keys table, as the xpubs can be stored
directly on the relation. In turn, this removes the need for loading key
material with either a query or subquery.

This does come at the cost of losing normalization of the keys, as one
key is stored in multiple places. The created_at timestamp per key is
also lost, so it cannot be determined the first time a key was added to
the system easily. Since the xpub can act as an ID, the penalty of
denormalization is minimized.

Closes chain/chainprv#107.
Reviewers: @jeffomatic