api: cache id/secret authentication results
The process of authenticating tokens has a significant runtime cost.
Particularly, bcrypt.CompareHashAndPassword takes ~22ms. Loading secrets
and user ids from the database also has a cost in the millisecond range.

Since every request must be authenticated, there is a large performance
improvement by caching the results of the authentication process. The
results are tied together with an expiration, at which point the process
must be run again.

Closes chain/chainprv#111.
Reviewers: @ryandotsmith