api/appdb: use a pg sequence for next addr index
There was too much contention updating the bucket record
every time we create an address. Incrementing a sequence
is cheap and fast. This means that any bucket will have
gaps in its sequence of addresses, possibly large gaps,
but that's ok.

We'll have to be careful if we ever shard this database,
to ensure that any given bucket doesn't repeat old addr
indexes. However, if we shard by bucket id, that should
not be much of a problem.

Closes chain/chainprv#118.
Reviewers: @ryandotsmith