api/utxodb: new package
We had a plv8 function to reserve outputs, but it was
not fast enough. We've replaced it with in-memory lists
of outputs and reservations.

Caveats

By special request, this patch doesn't add any new
tests, and it deletes some tests that apply to code
that was replaced with equivalent code in the new
package. Soon we should resurrect the deleted tests and
translate them to apply to the new logic, plus we should
add new tests for the wholly new logic.

We currently don't save short (<2 minute) reservations
at all, so restarting the process would cause us to lose
that information. We plan to begin saving all
reservations again, but in a separate reservations
table, not in the outputs table, again for performance.
SQL updates on a large, high-velocity table are
generally much slower than inserts and deletes, and this
situation is no exception.

Closes chain/chainprv#181.
Reviewers: @erykwalder