api/appdb: allocate address indexes in blocks
Creating a new address used to take two database round
trips: one to load the wallet/bucket info plus the next
address index within the bucket, followed by computing
the keys and address in memory based on the derivation
path, then a second round trip to insert the record.

Now we cache the wallet/bucket info, and reserve blocks
of 10,000 address indexes at a time, which we can
distribute from memory. On average, this eliminates the
first round trip.

Closes chain/chainprv#184.
Reviewers: @erykwalder