api/utxodb: fix panic in Reserve
Reserve collects a list of pointers to UTXO objects,
then, when an error occurs, it tries to "unreserve"
them, which consists of updating the timestamp and
fixing the heap organization. These two steps are each
done while holding a lock, but other things can happen
in between. In particular, another goroutine can remove
items from the pool. When the first goroutine tries to
unreserve the items, it assumes the UTXO objects it
holds pointers to are still in the heap. If they've
since been removed, unreserve can panic.

The fix is to check whether a UTXO is actually in the
heap when we want to unreserve it. If it's been deleted,
unreserve should be a no-op for that UTXO.

Closes chain/chainprv#187.
Reviewers: @jeffomatic