api/utxodb: fix crash deleting missing utxos
We weren't properly restoring reserved utxos when
pool.reserve returned an error. Returning nil for the
result set overwrote the return parameter utxos, causing
us to lose the list of UTXO objects that need to be put
back in the heap.

Thanks @jeffomatic for tracking this down.

The fix is to stop using labeled return parameters.
Returning nil will then have no effect on the value of
the variable 'utxos'.

Closes chain/chainprv#188.
Reviewers: @jeffomatic