api/assetdb: regenerate schema
The generated schema.sql.go needs to be kept in
sync so chainbot can use it when it provisions a new
stack. It didn't happen this time because we were adding
this new file concurrently with other branches updating
schema.sql. In the future we should run 'go generate' in
the same commit as any changes to schema.sql.

Closes chain/chainprv#202.
Reviewers: @bobg