api/utxodb: fix a race that causes duplicate utxos
Consider the following sequence:

1. Goroutine 1 calls ApplyTx, then insert, to insert
utxos into the database and the pool, but does not yet
commit itx db transaction.
2. Goroutine 2 calls pool.init, which throws away the
utxo list and loads an empty set of utxos from the db.
3. Goroutine 1 commits its db transaction.

Now we have valid utxos in the db, but an empty list in
memory. The fix is to check for dups in insert as well
as keeping the existing list in init.

Thanks to @erykwalder for the analysis.

Closes chain/chainprv#240.
Reviewers: @erykwalder