Add checkPasswordReset API call
This API call allows us to immediately show users an error message if
the password reset link they are using has already expired. Currently,
the user enters a new password and confirmation before we attempt to
verify the credentials in the password reset link. This leads to a poor
UX, since the password update form delivers the impression that the link
is valid.

Closes chain/chainprv#241.
Reviewers: @erykwalder