cmd/chainbot: fix server config option syntax
The parameter ssl_renegotiation_limit is a "server
config" option, defined in Chapter 18 of the postgres
9.4 manual:
http://www.postgresql.org/docs/9.4/static/runtime-config.html

Server config options are specified with the "options"
query parameter. See
http://www.postgresql.org/docs/9.4/static/libpq-connect.html#LIBPQ-CONNECT-OPTIONS

Note that setting the options value to "-c foo=bar" as
described in the manual requires escaping the space and
equals sign, which results in "options=-c%20foo%3Dbar".

Closes chain/chainprv#250.
Reviewers: @jeffomatic