fedchain/script: remove type Script
Package database/sql has rules for converting from one
type to another that are a bit subtle. If you give Scan
the exact type *[]byte, then it will copy bytes from the
database into a new slice for you; if you give Scan the
exact type *sql.RawBytes, then it will let you borrow a
buffer that will be recycled by package sql. These two
cases are documented explicitly. However, if you declare
your own type T []byte and give Scan a *T, then the type
conversion routine will fall through to generic
reflection code, that checks whether the referent of
the destination pointer is compatible with the raw
source value from the db. In this case it checks whether
T is compatible with []byte, which it is. In that case,
it assigns the value directly (using reflect) which
produces the same behavior as RawBytes.

See https://github.com/golang/go/blob/release-branch.go1.5/src/database/sql/convert.go#L206-L209.

This pitfall can be worked around by converting the
pointer in the call, like Scan((*[]byte)(&t)), but it
seems safer to avoid creating such types as T in the
first place, for values that we want to scan from the
database.

This type Script was originally made defensively, so
that we wouldn't have to worry about mixing up tx
scripts with other byte slices. However, we've used
[]byte before for this purpose in other codebases (in
papi and other bitcoin work) without encountering the
hypothetical problems this type was meant to protect
against. The btcsuite packages use []byte for scripts
throughout. So it seems like a simple and safe solution
to go back to using []byte for tx scripts.

Yet another alternative is to keep our special Script
type, but make sure it's not assignment-compatible with
[]byte, by making it something like struct { v []byte },
and writing methods Scan and Value. If using []byte
becomes a problem, we can consider this option.

Closes chain/chainprv#321.
Reviewers: @erykwalder