log: format stack on multiple lines
Splunk is configured (by default, and in our live
system) to collect multiple lines into a single log
event. It requires each event to start with a line
containing a well-formed timestamp. Subsequent lines
without a timestamp will be grouped with that first
line.

We can format stack traces in the natural way over
multiple lines following a normal log message, and this
will be interpreted by Splunk as a single event.

Closes chain/chainprv#323.
Reviewers: @ryandotsmith