api/asset: don't use a dbtx for block storage
Having this long running transaction slows down other processes as they
wait for this transaction to complete.

Having the blocks table is the canonical representation of the
blockchain's state, and all other tables can be derived from this one
table if necessary. The current architecture is not designed around
this, but can be. It is also not necessary for all the pieces to show up
in the database atomically, for example, we can delete outputs consumed
by a block before inserting newly created outputs.

Closes chain/chainprv#343.
Reviewers: @erykwalder