api/asset: store only new transactions in a block
Most transactions in each block will already be in our
database because the process has already seen them. For
the sandbox, this is because it finalized them to begin
with and inserted them into the pool and the historical
txs table. For a production configuration, most
transactions not finalized by the current process will
have been over the gossip protocol into the pool.

Closes chain/chainprv#357.
Reviewers: @erykwalder