migrations: fix date
We want to treat branch master as a linear history, and
have migration files in that linear history always added
in order. This makes it easier to find the set of
migrations that need to be run when we deploy new code
to each stack.

So rename this file to reflect when it was merged.

Closes chain/chainprv#384.
Reviewers: @jeffomatic