api/txdb: don't apply utxo to reserver twice
This commit fixes another race when applying a block
containing an output and spending that output
concurrently.

1. build+finalize tx0 (containing output tx0.0)
2. generate block containing tx0
3. insert tx0.0 into blockchain
   - side effect: removes tx0.0 from pool
   - so tx0.0 is part of block's reserver delta
4. build+finalize tx1, spending tx0.0
5. tx1 deletes tx0.0 from reserver
6. block-applying goroutine applies its delta
   to reserver, inserting tx0.0 (wrong!)
7. double spend tx0.0 can happen now.

The fix is to keep track of which outputs were in the
pool as we update them. Prior to
a5e7bd8efb110cfa8be2c6d2f62f0e34deef542f, this was a
consequence of keeping the blockchain separate from the
pool: we would insert outputs into the blockchain, and
then query the contents of the pool to see which ones
are present, then finally delete items from the pool;
this required outputs to exist in both the blockchain
and the pool concurrently. Now, we effectively move
outputs atomically from pool to blockchain (by updating
confirmed=t), so we also remember the result of that
update to keep track of which outputs were in the pool.

Credit to @jeffomatic for predicting this and to
@erykwalder for analyzing it once it actually happened.

Closes chain/chainprv#407.
Reviewers: @erykwalder