bin/git-land: use PR description as commit message
When git-land lands a PR, it used to use the message from the first
commit on the branch as the message for the final commit. It would
"fixup" all the subsequent commits and discard their messages. It then
opened an editor to allow editing of the final message before pushing
that commit onto the master branch. However, this was slightly error
prone: any necessary changes to the commit message would have to be made
after the review process, so they could not be reviewed. It was easy to
make a mistake or simply forget to make changes altogether.

Now, git-land gets the commit message from the pull request title and
body (the body is essentially the first comment). These fields are
usually populated from the first commit's message, so they start with
the same contents as before. However, we can edit them in the pull
request UI during the review process, and git-land will use them as the
commit message.

If this is successful, we might go on to have git-land skip opening the
editor on the commit message entirely, since it is expected to have
already been edited and reviewed in the pull request UI.

Closes chain/chainprv#459.
Reviewers: @jeffomatic