api/appdb: define key-index funcs in the schema
This was over-engineered. It is nice to have functions
loaded automatically into the schema when the app
starts, and it may have been worthwhile when we had many
complicated stored procedures. But the mechanism we use
to achieve this, setting the default postgres schema
search path, would be more useful to achieve other
things, such as logical sharding and/or logical
partitions (e.g. distinguishing blockchain state from
local manager state).

It's much simpler to define functions in the same place
as all the rest of the SQL.

Closes chain/chainprv#467.
Reviewers: @bobg