api: remove list of papi errors in comments
The list of old papi errors was useful for ensuring compatibility with
existing clients, but enough time has passed that it doesn't seem
relevant anymore. It seems like some error codes in the new API were
being chosen to ignore collisions with the list of papi errors, which
seems unnecessary.

Closes chain/chainprv#470.
Reviewers: @kr