log: print stack, if present, for every error message
Previously, function Error would extract the stack from
the error value, and add it to the Write call
as an explicit kv parameter under KeyStack. But there
are places (for example, api/http.go in logHTTPError)
where we call Write directly, providing a non-nil error.
This changes Write to look for the stack in error
parameters, and simplifies Error.

Closes chain/chainprv#476.
Reviewers: @jeffomatic @kr