fedchain/bc: Use spec defined signature hashes
The fedchain spec uses a signature format that involves hashing the
inputs and outputs separately, then combining those with other
transaction data to generate one final hash. Notably, this hash does not
include any subscript which must be signed, which was previously either
the redeem script of a p2sh addr, or the pk script of any other addrs.

The hash does now include the asset amount of the previous utxo, which
allows clients to safely sign inputs without needing all of the previous
transactions to determine that they are not losing funds.

Closes chain/chainprv#492.
Reviewers: @kr