bin/git-land: fix message editing flow
The intention behind chain/chainprv#459 was to change how we compose
the default message, but to preserve the step that opens
the editor to allow last-minute edits. It inadvertently
also removed the editor step.

We might still want to make that change deliberately at
some point, but for now we should stick with the
existing flow until we have experience under our belt
that indicates we essentially never want to make those
last-minute edits.

Closes chain/chainprv#493.
Reviewers: @erykwalder