crypto/hsm/thales/secp256k1: build rules for secp256k1
This directory is a C module to compile libsecp256k1 for
the Thales HSM Secure Execution Environment.

The original library from
https://github.com/bitcoin/secp256k1 was included
unmodified in chain/vendor/secp256k1 in a previous
commit. This new directory contains everything necessary
to build it for a Thales HSM.

Note that it can only build on a system with the Thales
cross-compile toolchain installed.

Closes chain/chainprv#505.
Reviewers: @bobg