gitignore: allow Makefile in subdirectories
Rules in .gitignore apply to all subdirectories unless
they begin with a / char, in which case they are
interpreted relative to the .gitignore file itself.

This change updates .gitignore to ignore Makefile only
in the root of the repo, and not ignore it elsewhere.
While we're at it, make the other rule ignoring editor
backup files ending with ~ simpler.

Closes chain/chainprv#501.
Reviewers: @erykwalder