Readme: new env var CHAIN, for root of repo
As we include code in more languages with different
build systems, it'll be useful to have a way to refer to
the root of the repo from anywhere. We have the user set
an environment variable, and we can use it as $CHAIN
in shell scripts, makefiles, and other automation.

Closes chain/chainprv#506.
Reviewers: @erykwalder