vendor/secp256k1: update upstream dependency
The current commit on our fork is
1a8375ccee4909b5d4d7c5c3b124e7744e73c5aa. Its tree is
8bac7836c4a526be64dd185cc9bff71b3a554bee. This commit
includes that tree as a subtree with prefix
vendor/secp256k1.

Our fork:
https://github.com/kr/secp256k1

Upstream pull request:
https://github.com/bitcoin/secp256k1/pull/377

Original commit message:

Put precomputed ecmult table in build dir

Include ecmult_static_context.h using the include path
so that it can be generated with a VPATH-style build in
a separate directory, when the original source directory
is not writable.

Closes chain/chainprv#525.
Reviewers: @erykwalder