crypto/hsm/thales/secp256k1: update build rules
Experiments indicate that the total stack space
available to an SEE module is about 32kB. One of the
setup functions in this library,
secp256k1_ecmult_gen_context_init (called by the public
entry point secp256k1_ecmult_context_build), was
allocating more than this on the stack, causing the
module to crash.

So change the rules (taking advantage of the patch to
libsecp256k1 in dec413682d23432b48dea0faea05990cb7d3bd23)
to build with USE_ECMULT_STATIC_PRECOMPUTATION. This
avoids the large stack allocation and the setup function
completes successfully.

Closes chain/chainprv#528.
Reviewers: @erykwalder