api: ensure issuance sigscript is generated correctly
Updates to the transaction template format silently broke the correct
generation of signature scripts for issuance transactions. Because we
are not currently validating issuance transaction scripts, this did not
break the finalization of issuances.

However, this meant that the correct asset ID could not be extracted
from the issuance signature script, and as a result, asset definitions
were not being indexed correctly during block application.

To prevent future confusion, this commit also renames the transaction
template input property from "RedeemScript" to "SigScriptSuffix", to
properly identify the role of the field. The serialized key remains
"redeem_script" to ensure backward compatibility with the current
version of the Java SDK.

The origin of this bug appears to be cccf88e, when we introduced a
slight tweak to the transaction template input format to support
orderbook contracts.

Closes chain/chainprv#547.
Reviewers: @erykwalder