api: add batch asset query to auditor node
This adds a new call, POST auditor/get-assets, which returns results for
multiple assets in a single request. This is useful for driving
interfaces which receive asset IDs in bulk, such as in balance queries.

To enable this, appdb.GetAssets() and txdb.AssetDefinitions() have been
added, and the singular versions of those calls have been re-implemented
using these new functions.

The new batch endpoint preserves the buggy semantics of the existing
GET auditor/assets/:asset_id call, which returns results for assets
that have not yet been issued. This will be fixed in a later PR.

Closes chain/chainprv#552.
Reviewers: @bobg