api: fix response for orderbook find API
This commit patches the orderbook "find" endpoints to include the
seller script, which is necessary to complete orderbook sales without
forcing the user to parse the PK script.

The changes here are not architecturally ideal, but they are minimally
intrusive while fixing the use case for end-users. A proper refactor is
probably due at some point in the future.

Closes chain/chainprv#565.
Reviewers: @bobg