api/txbuilder: include AssetAmount in template
Include the AssetAmount data in the template, so that we can
compute the signature data without hitting the db again to load
this info. This will also be used by the client, eventually, to
compute the signature data itself (to avoid trusting the server
to compute it correctly).

This will require SDK updates (at least in Java) so they preserve
these fields in later API calls.
See https://github.com/chain-engineering/chain-sdk-java/pull/63.

Closes chain/chainprv#589.
Reviewers: @erykwalder