fedchain/patricia: implement patricia merkle tree
Implements the patricia merkle tree used by the fedchain spec for
computing the root hash of the state. A patricia merkle tree is just
like a regular patricia/radix tree with the addition of a merkle hashing
structure. This implementation uses a binary patricia tree.

Patricia trees use keys to determine the placement of a value
within the tree. Each bit of the key is used in order as directions for
how to navigate down the tree. Only as much of the key is needed to
identify a unique path is used when inserting and creating branches.

A merkle tree is just a tree in which the hash of a parent node is
determined by the hash of its children. If a parent has multiple
children, then its hash is made by hashing the hashes of its children.
If it only has one child, then its hash is simply its child hash, with
no additional hashing taking place.

Closes chain/chainprv#616.
Reviewers: @kr