crypto/hsm/thales: command protocol spec
We've gotten the "hello world" up and running: a client program
and an SEE firmware module that talk to each other but don't do
anything interesting. The next step—this patch—is to design a
protocol to communicate all the functions they'll need to
perform. After this, we'll implement the spec and have a fully
functional signing module, and eventually connect it all the way
through the SDK into applications.

Closes chain/chainprv#633.
Reviewers: @bobg