fedchain/bc: remove tx.IsIssuance func
This function was previously used to determine if a transaction was an
issuance transaction. Soon, issuance and transfer inputs will be used
inside of a single transaction, so it is incorrect to call a transaction
an "issuance transaction."

There were multiple calling sites that assumed an issuance transaction
had no transfer inputs, and all of those assumptions have been corrected
to account for the mingling.

There were two big call site changes:

First, to the transaction validation code. Balancing the inputs and
outputs has to account for issuance inputs, and issuance inputs need to
be skipped individually for script checking.

Second, calculating the amount issued for transactions now needs to know
about the AssetAmount of inputs consuming utxos. This required changes
to both the fedchain and txdb packages, as well as other sites that
called functions within these packages.

Closes chain/chainprv#650.
Reviewers: @bobg