fedchain: introduce memStore to simplify tests
Rather than having the fedchain package rely on outside packages and
databases for tests, it is more ideal to have it be as self-contained as
possible. This removes the need for external pg and fixture packages. As
a bonus, assettest is now only used within the api package and its
subpackages.

Closes chain/chainprv#661.
Reviewers: @bobg @kr