fedchain/bc: remove AssetDefinitionPointer struct
This struct contained two fields, the asset id and the definition
pointer (hash of the definition). The asset id piece of this was
rarely used, since often these pointers are stored in maps which use the
asset id as a key. This change required a minor change to the
SaveAssetDefinition state.ViewWriter function.

Closes chain/chainprv#656.
Reviewers: @kr