fedchain: move database transactions to store
To support multiple storage mechanisms, the fedchain package should not
make any assumptions about the nature of the underlying store. It was
previously assumed that every store was using a postgres connection.

Although some queries are no longer being executed inside of the
transaction, there should not be any harm in this, since they were
select queries and therefore added no safety.

Closes chain/chainprv#657.
Reviewers: @kr