net/http: remove ListenAndServeTLS
Previously, cmd/api.main() called our own version of the
http.ListenAndServeTLS function because we wanted to pass the
key pair in directly (instead of as filenames for the location of the
certificates).

If we call http.Server.ListenAndServeTLS instead, we don't need our
own version of ListenAndServeTLS and don't need to maintain code copied
from the standard library's net/http package. Server's ListenAndServeTLS
method will not load the certificates if there are already certificates
on the server's TLSConfig and the provided filenames are empty strings:

https://golang.org/src/net/http/server.go?s=67534:67602#L2235

Note that the tcpKeepAliveListener removed in this commit was also copied
from net/http. The http.Server.ListenAndServeTLS function will give us
equivalent functionality.

Closes chain/chainprv#672.
Reviewers: @kr