fedchain/txscript: add execution stack
This commit refactors the txscript engine to introduce the concept of an
execution stack. This stack consists of stack frames, which include
a script and a program counter for that script. A significant amount of
state that used to live on the Engine has been moved to the stack frame:

  * the program counter
  * the script version value was previously reset on each new script, so
    I moved it onto the stack frame.
  * conditional stacks were expected to be empty between scripts to
    prevent conditionals spanning pkscript and sigscript. This was moved
    onto the stack frame.

The OP_EVAL opcode implementation also changed in this commit. Instead
of modifying the script in-place and inlining the eval'd script, the
OP_EVAL opcode will push a new stack frame with the eval'd script. This
does change the behavior slightly in that eval'd scripts can now provide
a script version number. Previously, an eval'd script would _ALWAYS_ be
interpreted using the caller script's version number.

Additionally, previously the max opcode limit was enforced for pkscript
and sigscript separately. Now the max opcode limit is enforced across
the entire script execution, including pkscript, sigscript and eval
scripts. A few tests that relied on the old behavior were removed.

These changes are intended to pave the way for WHILE... ENDWHILE loops.
https://github.com/chain-engineering/fedchain/blob/master/documentation/p2c/loops.md#option-2-whileend-with-inline-body

Closes chain/chainprv#687.
Reviewers: @bobg