crypto/hsm/thales: key generation
This change implements key generation, including the Java package
functions to use the HSM key generation operation, set an
appopriate ACL, and tell the firmware to load the new key; plus
the firmware code to retrieve the key material and verify that
it's valid. This process is described in more detail in the
Readme.

The other functions, signing and derivation, are still to come.

This commit also updates vendor/ct to the current upstream commit
at https://github.com/kr/ct. The current commit on upstream
branch master is 871416d88a, which includes subdirectory ct as
tree c3e9097ca5. This commit includes that tree as a subtree at
vendor/ct.

Closes chain/chainprv#685.
Reviewers: @bobg