cmd/chainbot: do not write db schema on provision
Since the introduction of the migratedb tool, it is not correct to use
schema.sql to initialize databases except for unit tests. This commit
updates chainbot's provision command so it no longer writes the database
schema. As a consequence, chainbot no longer initializes a default API
user account (arguably, this has not been a useful feature to begin with).

Readme.md has been updated with a step-by-step guide on how to provision
new stacks, along with a TODO with ideas on how to further automate the
provisioning process.

Closes chain/chainprv#692.
Reviewers: @ryandotsmith