fedchain: track state merkle root
This update begins generating and validating the state merkle root of
blocks. This involves tracking all of the available utxos, as well as
current asset definition pointers. This state hash can be used in the
future to provide snapshots of the network without having to validate
the entire chain. It can also be used to provide merkle proofs about the
current circulation of asset, or the existence of a given utxo.

Closes chain/chainprv#746.
Reviewers: @kr