api: add hack feature flag for cross-project transfer
This commit adds an optional environment flag,
ENABLE_CROSS_PROJECT_XFER_HACK, that allows API users to execute
build-transaction API calls across project lines using account IDs. This
removes the need to implement an address exchange flow in the
application itself. Whether this accurately represents real-world
organizational borders depends on the application, so this is
implemented as a configurable flag.

This feature should considered a hack, since it exposes a flaw in our
current sandbox authorization model. In particular, it is not clear
whether a project should be used to model a single "core" (i.e., a
fully independent entity on the blockchain), or simply provide
rudimentary authorization between nodes within the same core (where the
entire sandbox represents a single core). In principle, only the latter
case allows for transfers between accounts in different projects.

As of right now, the most robust vision for redesigning the sandbox is
as follows:

1. Cores (independent entities on a blockchain) map to sandbox deploy
targets on a one-to-one basis. Cross-organization interaction on a
single blockchain will be modeled in a true-to-life fashion: multiple
sandbox targets will be deployed and configured to communicate with each
other.
2. Authorization within cores will be simplified to admin/member;
projects will cease to exist. Visibility into the contents of separate
"services" (manager and issuer nodes) within the core must be handled by
the application, not the core.

Closes chain/chainprv#732.
Reviewers: @bobg