api/smartcontracts/voting: new package
This commit adds a single clause of the voting rights contract. The
transfer clause moves a voting right token without extending the chain
of ownership (for key rotation, moving within accounts of same org,
etc).

This is the first contract that requires knowledge of the destination in
the sigscript AND knowledge of the source in the destination's pkscript.
For now, this special-cases the parsing of build request vrtoken sources
and destinations. This can be replaced with a better abstraction once
we have a better idea of what that should be.

Closes chain/chainprv#735.
Reviewers: @bobg