crypto/hsm/thales/xprvseemodule: fix cross-compile build
This change fixes the build using the cross-compile
toolchain. The file op.c isn't built as part of the unit
tests because it uses libraries that are only available
in the HSM's toolchain.

I also snuck in a bug fix in the error case when the key
table is full, and a tweak to rebuild the generated test
vector C code whenever gen.go changes.

Reviewed at https://rbcommons.com/s/chain/r/2/
Closes chain/chainprv#739
Reviewers: @bobg