api/appdb: don't delete assets during blockchain reset
Currently, we truncate the assets table when resetting the blockchain.
According to the spec, this is the correct thing to do, since asset IDs
are generated using the hash of the genesis block, and the genesis block
theoretically changes between blockchain resets.

However, deleting assets is a significant nuisance for users. What's
more, we have so far not been mixing the genesis block hash into asset IDs.
It thus seems preferable to preserve assets between blockchain resets.

In the future, it's likely that we'll begin to include the genesis block
hash into asset IDs. We should be able to dodge around this issue by
preserving the genesis block across resets (rewinding the blockchain to
height 1 instead of height 0), only deleting assets when the user
requires a new genesis block.

Closes chain/chainprv#743.
Reviewers: @kr