database/pg/pgtest: create a new db for every test
Using database transactions for isolating tests from each other
had some drawbacks. Inside a transaction, it's not possible to
perform concurrent database queries, and we occasionally have the
need to test concurrent code. Also, we often have the need to
test code that wants to start and stop a database transaction.
Previously, we accomodated this with an interface that presented
a fake transaction to the code under test, while the real
transaction had already been created by the test function.

This commit changes the mechanism for test isolation. Now we
create an entire new database (using 'CREATE DATABASE ...') for
every test that needs database access. This lets the code under
test have free rein to do whatever it needs inside that database,
including starting and stopping multiple transactions, rolling
back, executing concurrent queries, and even operating on
multiple schemas.

This also obviates the need for the SchemaDriver, which helped
isolate tests in one package from those in another.

While we're at it, we also streamline the pgtest interface a bit.

However, there is one big drawback to this change: our tests are
now much slower. This is unfortunate but necessary. "Make it
right before you make it fast. Keep it right when you make it
fast." It wasn't right; now it is.

Future work: there is now code in chain/database/pg that is
unneeded, and it should be deleted, including the transaction
mocking functions and SchemaDriver mentioned above. We also want
to change database-transaction-using code to require the Tx
object as explicit function parameters, using static type
checking, rather than pulling it out of the context with a
runtime assertion. The above changes will make this simpler.
Finally, we plan to add back in the option to use db
transactions to isolate those tests that can run in
exactly one transaction. That should regain a good chunk
of the performance lost in this patch.

Closes chain/chainprv#866.
Reviewers: @bobg