api/smartcontracts/voting: basic issuance
We're punting on the voting right issuance contract, but for the time being
we need a way to create voting right assets. This commit allows transactions
to send a single unit of an arbitrary asset into the voting right contract
in order to create a vrtoken. It doesn't enforce any of the contractual
guarantees of the issuance script (ex, circulation = 1, executing the
issuer/admin script, etc)

Closes chain/chainprv#756.
Reviewers: @jeffomatic