api/txbuilder: structure sigscript in template
Some voting right clauses require signatures from two parties: the current holder and the voting system administrator. Both the current holder and the admin are represented as arbitrary scripts (they can be addresses, multisig scripts, etc). In order to support them, we'll need to add structure to the signature scripts in transaction templates to prevent the interweaving of redeem scripts and signatures.

This commit adds a new transaction template input field `signature_components` to replace `signatures` and `sigscript_suffix`. For now, it's only used for the voting system inputs. We can update the SDKs to understand the new format and then move other inputs over to the new format.

See discussion in #devlog https://chainhq.slack.com/archives/devlog/p1460747025002467

Closes chain/chainprv#800.
Reviewers: @erykwalder