api/smartcontracts/voting: Add finish clause
Adds clause for marking a voting token as being in the finished state.

Also, modifies the `stateFinished` constant to be `0x10`. Currently, our scripting language doesn't have support for bitwise opcodes, so we'll need to use `ADD` and `SUB` for setting the finished bit. The previous constant `0xF0` included the sign bit which made that a little more difficult.

Closes chain/chainprv#828.
Reviewers: @kr