cos/bc: add serialization flags to tx
This is a spec update. Each transaction now starts with
one byte of serialization flags. Three are defined, the
rest are reserved for future use and must be 0. In
addition, the current implementation rejects any flag
set other than 0x7 when reading a transaction.

See https://git.io/serflags for more info.

Closes chain/chainprv#831.
Reviewers: @jbowens