cos/bc: remove external test dependency
Using the 'package foo_test' form can be inconvenient
and confusing. We should not go too overboard.

In this case, package bc, which is concerned with the
wire format of blockchain data structures, shouldn't
depend on txscript, which provides script execution
semantics. There's no need to run scripts in order to
serialize and deserialize transactions. The contents of
the scripts don't matter.

Closes chain/chainprv#826.
Reviewers: @jbowens