iago: wait between retries, retry backup more
This adds a wait between retries, and increases the number of times that we attempt to pg backup from an upstream. This specifically targets the problem where syncs and asyncs would crash because they were attempting to backup from an upstream that hadn't started its postgres processes yet. (30 retries x time.Second wait period = 30 seconds for the backup to succeed. Perhaps gratuitous.)

Closes chain/chainprv#840.
Reviewers: @kr