api: Revert waiting for tx confirmation in TestTransfer
As described in https://chainhq.slack.com/archives/devlog/p1462225961001161:

> we're using the same `ctx` object in the test function and in the block-generat
ing goroutine it spawns. this means that the outer function can create a savepoint, then the inner function creates a savepoint, then the outer function commits the savepoint, then the inner function tries to commit but because it was nested, the savepoint no longer exists. also, if the goroutine uses a separate db connection, its db changes are not visible to the test function (because the test function runs inside its own disposable db transaction). any bright ideas for working around this? maybe just ditch the goroutine...?

So, ditching the goroutine for now until we come up with a non-db-transaction-based way to isolate tests.

Closes chain/chainprv#847.
Reviewers: @kr